+++ API/MarketAPI/Data/Models/Order.cs
@@ -16,9 +16,9 @@ namespace MarketAPI.Data.Models
         public bool IsDenied { get; set; } = false;
 
         [Required]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
         [Required]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [Required]
         public string Address { get; set; }
