+++ API/MarketAPI/Models/Common/Email/Models/ConfirmationEmailModel.cs
@@ -5,7 +5,7 @@ namespace MarketAPI.Models.Common.Email.Models
 {
     public class ConfirmationEmailModel
     {
-        public ConfirmationEmailModel(string email, List<ConfirmationEmailOrderDTO> orders, double price, BillingDetailsDTO billingDetails, int year, string userName)
+        public ConfirmationEmailModel(string email, List<ConfirmationEmailOrderDTO> orders, decimal price, BillingDetailsDTO billingDetails, int year, string userName)
         {
             Email = email;
             Orders = orders;
@@ -17,7 +17,7 @@ namespace MarketAPI.Models.Common.Email.Models
         public string UserName { get; set; }
         public string Email { get; set; }
         public List<ConfirmationEmailOrderDTO> Orders { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public BillingDetailsDTO BillingDetails { get; set; }
         public int Year { get; set; }
     }
