+++ MarketPortal/Market/Services/Inventory/IInventoryService.cs
@@ -7,7 +7,7 @@ namespace Market.Services.Inventory
     {
         public Task AddStockAsync(StockViewModel model);
         public Task UpStockAsync(ChangeStockViewModel model);
-        public Task DownStockAsync(int id, double quantity);
+        public Task DownStockAsync(int id, decimal quantity);
         public Task DeleteStockAsync(int id);
         public Task<List<Stock>> GetSellerStocksAsync();
     }
